# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "latihanApi"
  version: "0.6.0"
  description: "Personal sessions, plans, profiles and measurement history. /exercises has been removed. Lists use cursor pagination (limit, cursor, next_cursor). Requests are rate limited per client address, per account, and more strictly on /api/v1/auth; limited requests get 429 with Retry-After. Every /api/v1 operation except /api/v1/auth/* requires a bearer token from the configured provider. The /api/v1/auth endpoints and /.well-known/jwks.json exist only with the built-in jwt provider. POST /api/v1/users links the caller's identity to a new profile; other operations require that profile. Callers only ever see and change their own records: owners come from the token, never from input, and other users' records return 404. All responses include X-Request-ID. Input bodies are limited to 1 MiB; unknown JSON fields, unknown/repeated query parameters and multiple JSON values are rejected. Strings are trimmed. PUT fully replaces editable fields; no upserts. UUIDs and creation timestamps are server-managed. HTTP method errors use the same Error schema."
servers:
  -
    url: "http://localhost:8080"
security:
  - bearerAuth: []
...
paths:
...
  "/api/v1/sessions":
    post:
      operationId: "createSession"
      tags:
        - "Sessions"
      summary: "Create session"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/SessionInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "201":
          description: "Created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Session"
          headers:
            Location:
              schema:
                type: "string"
              description: "Relative URL of the created resource"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    get:
      operationId: "listSessions"
      tags:
        - "Sessions"
      summary: "List sessions"
      description: "The caller's sessions, newest performed_at first, then id DESC. No matches returns an empty items array. user_id is not accepted."
      parameters:
        -
          name: "limit"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 20
        -
          name: "cursor"
          in: "query"
          description: "next_cursor from the previous page; omit for the first page. Opaque."
          schema:
            type: "string"
            maxLength: 128
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Page"
          content:
            "application/json":
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "items"
                  - "limit"
                  - "next_cursor"
                properties:
                  items:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Session"
                  limit:
                    type: "integer"
                  next_cursor:
                    type: "string"
                    nullable: true
                    description: "Pass as cursor to get the next page; null on the last page."
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/sessions/{id}":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
          description: "Nonzero UUID"
    get:
      operationId: "getSession"
      tags:
        - "Sessions"
      summary: "Get session"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Found"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Session"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    put:
      operationId: "updateSession"
      tags:
        - "Sessions"
      summary: "Replace session"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/SessionInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Updated"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Session"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    delete:
      operationId: "deleteSession"
      tags:
        - "Sessions"
      summary: "Delete session"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "204":
          description: "Deleted, no body"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/plans":
    post:
      operationId: "createPlan"
      tags:
        - "Plans"
      summary: "Create plan"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PlanInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "201":
          description: "Created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Plan"
          headers:
            Location:
              schema:
                type: "string"
              description: "Relative URL of the created resource"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    get:
      operationId: "listPlans"
      tags:
        - "Plans"
      summary: "List plans"
      description: "The caller's plans, newest created_at first, then id DESC. No matches returns an empty items array. user_id is not accepted."
      parameters:
        -
          name: "limit"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 20
        -
          name: "cursor"
          in: "query"
          description: "next_cursor from the previous page; omit for the first page. Opaque."
          schema:
            type: "string"
            maxLength: 128
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Page"
          content:
            "application/json":
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "items"
                  - "limit"
                  - "next_cursor"
                properties:
                  items:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Plan"
                  limit:
                    type: "integer"
                  next_cursor:
                    type: "string"
                    nullable: true
                    description: "Pass as cursor to get the next page; null on the last page."
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/plans/{id}":
    parameters:
      -
        name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
          description: "Nonzero UUID"
    get:
      operationId: "getPlan"
      tags:
        - "Plans"
      summary: "Get plan"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Found"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Plan"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    put:
      operationId: "updatePlan"
      tags:
        - "Plans"
      summary: "Replace plan"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/PlanInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Updated"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Plan"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    delete:
      operationId: "deletePlan"
      tags:
        - "Plans"
      summary: "Delete plan"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "204":
          description: "Deleted, no body"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "409":
          description: "Dependent records prevent deletion."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/users":
    post:
      operationId: "createUser"
      tags:
        - "Users"
      summary: "Create user"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/UserInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "409":
          description: "This identity already has a profile."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "201":
          description: "Created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/User"
          headers:
            Location:
              schema:
                type: "string"
              description: "Relative URL of the created resource"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/users/{userID}":
    parameters:
      -
        name: "userID"
        in: "path"
        required: true
        description: "\"me\" or the caller's own user ID. Any other user ID returns 404."
        schema:
          type: "string"
    get:
      operationId: "getUser"
      tags:
        - "Users"
      summary: "Get user"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Found"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/User"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    put:
      operationId: "updateUser"
      tags:
        - "Users"
      summary: "Replace user"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/UserInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Updated"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/User"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    delete:
      operationId: "deleteUser"
      tags:
        - "Users"
      summary: "Delete user"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "204":
          description: "Deleted, no body"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "409":
          description: "Dependent records prevent deletion."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
  "/api/v1/users/{userID}/measurements":
    parameters:
      -
        name: "userID"
        in: "path"
        required: true
        description: "\"me\" or the caller's own user ID. Any other user ID returns 404."
        schema:
          type: "string"
    post:
      operationId: "createMeasurement"
      tags:
        - "Measurements"
      summary: "Create measurement"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/MeasurementInput"
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "201":
          description: "Created"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Measurement"
          headers:
            Location:
              schema:
                type: "string"
              description: "Relative URL of the created resource"
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "413":
          description: "Body exceeds 1 MiB."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "415":
          description: "Content-Type must be application/json."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
    get:
      operationId: "listMeasurements"
      tags:
        - "Measurements"
      summary: "List measurements"
      description: "Newest measured_at first, then id DESC. A missing user returns 404."
      parameters:
        -
          name: "limit"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 100
            default: 20
        -
          name: "cursor"
          in: "query"
          description: "next_cursor from the previous page; omit for the first page. Opaque."
          schema:
            type: "string"
            maxLength: 128
      responses:
        "429":
          "$ref": "#/components/responses/RateLimited"
        "401":
          "$ref": "#/components/responses/Unauthorized"
        "403":
          "$ref": "#/components/responses/ProfileRequired"
        "503":
          "$ref": "#/components/responses/ServiceUnavailable"
        "200":
          description: "Page"
          content:
            "application/json":
              schema:
                type: "object"
                additionalProperties: false
                required:
                  - "items"
                  - "limit"
                  - "next_cursor"
                properties:
                  items:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Measurement"
                  limit:
                    type: "integer"
                  next_cursor:
                    type: "string"
                    nullable: true
                    description: "Pass as cursor to get the next page; null on the last page."
        "400":
          description: "Invalid JSON, UUID, input, query, owner or plan reference."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "404":
          description: "Resource missing, belongs to another user, or session not attached to the requested plan."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
        "500":
          description: "Unexpected error; internal details are not exposed."
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Token from the configured AUTH_PROVIDER (Firebase ID token, Cognito access or ID token, or OIDC access token)."
  responses:
    RateLimited:
      description: "rate_limited: too many requests from this address or account. Retry-After gives the wait in seconds."
      headers:
        Retry-After:
          schema:
            type: "integer"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
...
    Unauthorized:
      description: "Missing or invalid bearer token. WWW-Authenticate describes the challenge."
      headers:
        WWW-Authenticate:
          schema:
            type: "string"
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
    ProfileRequired:
      description: "Authenticated, but the identity has no profile yet. Create one with POST /api/v1/users."
      content:
        "application/json":
          schema:
            "$ref": "#/components/schemas/Error"
...
  schemas:
...
    SessionInput:
      type: "object"
      additionalProperties: false
      required:
        - "name"
        - "performed_at"
        - "exercises"
      properties:
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        notes:
          type: "string"
          maxLength: 4000
          nullable: true
          default: ""
        performed_at:
          type: "string"
          format: "date-time"
        plan_id:
          type: "string"
          format: "uuid"
          description: "Nonzero UUID"
          nullable: true
        exercises:
          type: "array"
          minItems: 1
          maxItems: 100
          items:
            "$ref": "#/components/schemas/SessionExercise"
      description: "Full replacement. The owner is always the caller and is not part of the input. Optional plan_id must be one of the caller's plans. plan_exercise_id links each actual entry to an entry in its saved plan snapshot and must not repeat. Omitted plan_exercise_id means unplanned. Keeping the same plan_id retains the original snapshot; changing/reattaching captures the current plan; null/omitted plan_id detaches. Times are nonzero RFC3339."
...